@@ -5,11 +5,22 @@
   "scripts": {
     "serve": "vue-cli-service serve",
     "build": "vue-cli-service build",
-    "lint": "vue-cli-service lint"
+    "lint": "vue-cli-service lint",
+    "backend": "json-server --watch db.json"
   },
   "dependencies": {
-    "core-js": "^3.8.3",
-    "vue": "^2.6.14"
+    "@kyvg/vue3-notification": "^2.4.1",
+    "axios": "^0.27.2",
+    "bulma": "^0.9.4",
+    "core-js": "^3.25.2",
+    "json-server": "^0.17.0",
+    "velocity-animate": "^1.5.2",
+    "vue": "^3.2.13",
+    "vue-flash-message": "^0.7.2",
+    "vue-notification": "^1.3.20",
+    "vue-router": "^4.0.13",
+    "vuelidate": "^0.7.7",
+    "vuex": "^4.0.2"
   },
   "devDependencies": {
     "@babel/core": "^7.12.16",
